abstracting things out, moving things around
