+++ hourglass/views/main/__init__.py
@@ -1,5 +1,4 @@
 from flask import Blueprint
-from os.path import abspath, join, dirname
 
 
 main = Blueprint('main', __name__, template_folder='templates',
