+++ run.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python2
-import os
-from hourglass.command_line import run_hourglass
+from hourglass.shell import run_hourglass
 
 
 if __name__ == "__main__":
